edit: move toolbar modification to UIToolbar extension
